cardUpdate after any repository change
