+++ frontend/src/app/services/mat-listener.service.ts
@@ -1,18 +1,21 @@
 import { Injectable } from '@angular/core';
-import { Observable, Subject } from 'rxjs';
+import { BehaviorSubject, Observable } from 'rxjs';
 
 import { io } from 'socket.io-client';
 import { BACKEND_URL } from 'src/constants';
 import { Card } from '../models/card';
+import { GameStateService } from './game-state.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class MatListenerService {
   private readonly socket = io(BACKEND_URL);
-  private readonly cards: Subject<Card[]> = new Subject<Card[]>();
+  private readonly cards: BehaviorSubject<Card[]> = new BehaviorSubject<Card[]>(
+    []
+  );
 
-  constructor() {
+  constructor(private readonly gameStateService: GameStateService) {
     this.socket.on('connect', () => {
       console.log('Connected to server!');
     });
@@ -31,6 +34,11 @@ export class MatListenerService {
       // the reason of the error, for example "xhr poll error"
       console.log(err.message);
     });
+
+    // Get current list and use as initial value
+    this.gameStateService
+      .getAllCards()
+      .subscribe((cards) => this.cards.next(cards));
   }
 
   public getCards$(): Observable<Card[]> {
